Move quantization scripts to a scripts/ subdir.
