+++ invokeai/backend/quantization/scripts/load_flux_model_bnb_llm_int8.py
@@ -6,7 +6,7 @@ from flux.util import configs as flux_configs
 from safetensors.torch import load_file, save_file
 
 from invokeai.backend.quantization.bnb_llm_int8 import quantize_model_llm_int8
-from invokeai.backend.quantization.load_flux_model_bnb_nf4 import log_time
+from invokeai.backend.quantization.scripts.load_flux_model_bnb_nf4 import log_time
 
 
 def main():
